Fix animator not activating
