--- Assets/Scripts/SS3D/Systems/Entities/Humanoid/DummyRagdoll.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c820c500b2966a940907fe7f9855c107
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
